pyproject.toml
[tool.poetry]
name = "MelanomaGraphAnalysisV3"
version = "0.1.0"
description = "MelanomaBenignClassifier"
authors = ["Mohamed Abdelmoniem ma3546@fayoum.edu.eg"]

[tool.poetry.dependencies]
python = ">=3.11"
anthropic = ">=0.45.2"
joblib = ">=1.4.2"
matplotlib = ">=3.10.0"
networkx = ">=3.4.2"
numpy = ">=2.2.3"
pillow = ">=11.1.0"
scikit-image = ">=0.25.1"
scikit-learn = ">=1.6.1"
scipy = ">=1.15.1"
pandas = ">=2.2.3"

[[tool.poetry.packages]]
include = "src"